Split generic method template integrity tests
