+++ src/typechecker/tests/resolver_collection/function_method_templates/generic_methods/integrity.rs
@@ -1,0 +1,105 @@
+use super::*;
+
+#[test]
+fn collect_declarations_with_symbols_does_not_fallback_to_stale_ast_generic_method_template() {
+    let mut program = parse_program(
+        r#"
+Box: { value: i32 }
+Box.keep<T> = (self: Box, value: T) T { value }
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_return_type_for_test(Namespace::Value, "Box.keep", None);
+    if let Declaration::Method {
+        params,
+        return_type,
+        ..
+    } = &mut program.declarations[1]
+    {
+        params[1].ty = AstType::Named("Stale".to_string());
+        *return_type = Some(AstType::Named("AlsoStale".to_string()));
+    }
+    let mut tc = TypeChecker::new();
+
+    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
+
+    assert!(
+            !tc.generic_methods.contains_key("Box.keep"),
+            "resolver-backed collection should not keep AST-only generic method templates when resolver signature metadata is incomplete"
+        );
+}
+
+#[test]
+fn collect_declarations_with_symbols_clears_stale_generic_method_template_after_key_restore() {
+    let mut program = parse_program(
+        r#"
+Box: { value: i32 }
+Box.keep<T> = (self: Box, value: T) T { value }
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_return_type_for_test(Namespace::Value, "Box.keep", None);
+    if let Declaration::Method {
+        type_name,
+        method_name,
+        params,
+        return_type,
+        ..
+    } = &mut program.declarations[1]
+    {
+        *type_name = "Missing".to_string();
+        *method_name = "missing".to_string();
+        params[1].ty = AstType::Named("Stale".to_string());
+        *return_type = Some(AstType::Named("AlsoStale".to_string()));
+    }
+    let mut tc = TypeChecker::new();
+
+    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
+
+    assert!(
+            !tc.generic_methods.contains_key("Missing.missing"),
+            "resolver-backed collection should clear the stale AST generic method template key after resolver key restoration"
+        );
+    assert!(
+            !tc.generic_methods.contains_key("Box.keep"),
+            "resolver-backed collection should clear the restored generic method template key when resolver signature metadata is incomplete"
+        );
+}
+
+#[test]
+fn collect_declarations_with_symbols_uses_resolver_generic_method_name_for_body_type_refs() {
+    let mut program = parse_program(
+        r#"
+Box: { value: i32 }
+Box.keep<T> = (self: Box, value: T) T {
+    same: T = value
+    same
+}
+"#,
+    );
+    let symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    if let Declaration::Method {
+        method_name,
+        type_params,
+        ..
+    } = &mut program.declarations[1]
+    {
+        *method_name = "missing".to_string();
+        type_params[0].name = "Stale".to_string();
+    }
+    let mut tc = TypeChecker::new();
+
+    tc.collect_declarations_with_symbols(&program.declarations, &symbols);
+
+    assert!(
+            tc.diagnostics.is_empty(),
+            "resolver-restored generic method name and type parameters should avoid stale AST body type-ref diagnostics: {:?}",
+            tc.diagnostics
+        );
+}
